fixed 1.9 related issues | fixed unable to interact with entities | fixed help menu cmds showing twice

+++ src/bl4ckscor3/plugin/animalessentials/cmd/Kill.java
@@ -35,7 +35,7 @@ public class Kill implements IAECommand,Listener
 	public void exe(Plugin pl, CommandSender sender, Command cmd, String[] args) throws IOException
 	{
 		final Player p = (Player)sender;
-		
+
 		if(currentlyKilling.containsKey(p))
 		{
 			Utilities.sendChatMessage(p, "You can't kill multiple animals by issuing the command multiple times. Please use /()/ae kill <amount>()/ after rightclicking an animal or waiting.");
@@ -43,13 +43,13 @@ public class Kill implements IAECommand,Listener
 		}
 
 		int kills = args.length == 1 || Integer.parseInt(args[1]) <= 0 ? 1 : Integer.parseInt(args[1]);
-		
+
 		plugin = pl;
 		Utilities.sendChatMessage(p, "Please rightclick the animal you want to kill. " + ChatColor.RED + " THIS IS IRREVERSIBLE!!");
 		Utilities.sendChatMessage(p, "Kills available: " + kills);
 		currentlyKilling.put(p, new Killing(kills));
 		AECommands.setIssuingCmd(p, true);
-		
+
 		AbortRunnable task = new AbortRunnable(p);
 
 		task.runTaskLater(pl, 10 * 20L);
@@ -61,56 +61,62 @@ public class Kill implements IAECommand,Listener
 	{
 		if(currentlyKilling.containsKey(event.getPlayer()))
 		{
-			Entity entity = event.getRightClicked();
-
-			if(!Utilities.isAnimal(entity))
+			if(currentlyKilling.get(event.getPlayer()).enabled())
 			{
-				Utilities.sendChatMessage(event.getPlayer(), "You can't kill this mob, it's " + Utilities.aN(entity.getType().name(), false) + " /()" + (entity.getType().name() == null? "Player" : Utilities.capitalizeFirstLetter(entity.getType().name())) + "()/ and not an animal.");
+				Entity entity = event.getRightClicked();
+
+				if(!Utilities.isAnimal(entity))
+				{
+					Utilities.sendChatMessage(event.getPlayer(), "You can't kill this mob, it's " + Utilities.aN(entity.getType().name(), false) + " /()" + (entity.getType().name() == null? "Player" : Utilities.capitalizeFirstLetter(entity.getType().name())) + "()/ and not an animal.");
+					event.setCancelled(true);
+					return;
+				}
+
+				if(!event.getPlayer().hasPermission("aess.kill.bypass") && !Utilities.isOwnedBy(event.getPlayer(), entity, true))
+				{
+					Utilities.sendChatMessage(event.getPlayer(), "This is not your animal, you can't kill it.");
+					event.setCancelled(true);
+					return;
+				}
+				//x offset, y offset, z offset from the center, speed, amount, center, radius
+				ParticleEffect.SMOKE_NORMAL.display(0.0F, 0.0F, 0.0F, 0.5F, 100, entity.getLocation(), 255);
+				//Play the sound at the location
+				entity.getLocation().getWorld().playSound(entity.getLocation(), Sound.ENTITY_TNT_PRIMED, 1.0F, 1.0F);
+
+				entity.remove();
+				Utilities.sendChatMessage(event.getPlayer(), "Animal killed.");
+
+				if(currentlyKilling.get(event.getPlayer()).getAmount() == 1)
+				{
+					Utilities.sendChatMessage(event.getPlayer(), "Kills left: 0");
+					currentlyKilling.remove(event.getPlayer());
+					AECommands.setIssuingCmd(event.getPlayer(), false);
+				}
+				else
+				{
+					currentlyKilling.get(event.getPlayer()).decreaseAmount();
+					Utilities.sendChatMessage(event.getPlayer(), "Kills left: " + currentlyKilling.get(event.getPlayer()).getAmount());
+					return;
+				}
+
 				event.setCancelled(true);
-				return;
-			}
-
-			if(!event.getPlayer().hasPermission("aess.kill.bypass") && !Utilities.isOwnedBy(event.getPlayer(), entity, true))
-			{
-				Utilities.sendChatMessage(event.getPlayer(), "This is not your animal, you can't kill it.");
-				event.setCancelled(true);
-				return;
-			}
-			//x offset, y offset, z offset from the center, speed, amount, center, radius
-			ParticleEffect.SMOKE_NORMAL.display(0.0F, 0.0F, 0.0F, 0.5F, 100, entity.getLocation(), 255);
-			//Play the sound at the location
-			entity.getLocation().getWorld().playSound(entity.getLocation(), Sound.ENTITY_TNT_PRIMED, 1.0F, 1.0F);
-
-			entity.remove();
-			Utilities.sendChatMessage(event.getPlayer(), "Animal killed.");
-			
-			if(currentlyKilling.get(event.getPlayer()).getAmount() == 1)
-			{
-				Utilities.sendChatMessage(event.getPlayer(), "Kills left: 0");
-				currentlyKilling.remove(event.getPlayer());
-				AECommands.setIssuingCmd(event.getPlayer(), false);
+				Bukkit.getScheduler().cancelTask(taskIDs.get(event.getPlayer()));
+				taskIDs.remove(event.getPlayer());
 			}
 			else
-			{
-				currentlyKilling.get(event.getPlayer()).decreaseAmount();
-				Utilities.sendChatMessage(event.getPlayer(), "Kills left: " + currentlyKilling.get(event.getPlayer()).getAmount());
-			}
-
-			event.setCancelled(true);
-			Bukkit.getScheduler().cancelTask(taskIDs.get(event.getPlayer()));
-			taskIDs.remove(event.getPlayer());
+				currentlyKilling.get(event.getPlayer()).enable();
 		}
 	}
 
 	public class AbortRunnable extends BukkitRunnable implements BukkitTask
 	{
 		private Player p;
-		
+
 		public AbortRunnable(Player player)
 		{
 			p = player;
 		}
-		
+
 		@Override
 		public void run()
 		{
@@ -135,7 +141,7 @@ public class Kill implements IAECommand,Listener
 			return false;
 		}
 	}
-	
+
 	@Override
 	public String getAlias()
 	{
